% Generated by roxygen2 (4.0.1): do not edit by hand
\name{aggregation_code<-}
\alias{aggregation_code<-}
\title{Set aggregation code in a JAGS object}
\usage{
aggregation_code(object) <- value
}
\arguments{
\item{object}{a JAGS object.}

\item{value}{a character element or NULL.}
}
\value{
The replacement method changes the aggregation_code component of the object.
}
\description{
Sets the aggregation_code component of a JAGS object.
}
\seealso{
\code{\link{aggregation_code}} and \code{\link{jaggernaut}}
}

