\name{aggregation_code}
\alias{aggregation_code}
\title{Get aggregation code from a JAGS object}
\usage{
aggregation_code(object, ...)
}
\arguments{
  \item{object}{a JAGS object.}

  \item{...}{further arguments passed to or from other
  methods.}
}
\value{
The aggregation_code component of object.
}
\description{
Gets the aggregation_code component of a JAGS object.
aggregation code tracks scalar parameters.
}
\seealso{
\code{\link{jaggernaut}}
}

