\name{as.jags_data}
\alias{as.jags_data}
\title{Coerce to type JAGS data}
\usage{
  as.jags_data(x, ...)
}
\arguments{
  \item{x}{object to coerce.}

  \item{...}{further arguments passed to or from other
  methods.}
}
\value{
  If successful an object of class \code{jags_data}
  otherwise NA.
}
\description{
  Coerces to an object of type \code{jags_data}.
}
\seealso{
  \code{\link{jags_data}} and \code{\link{jaggernaut}}.
}

