\name{as.jags_data_frame}
\alias{as.jags_data_frame}
\title{Coerce to a JAGS data frame object}
\usage{
  as.jags_data_frame(x, ...)
}
\arguments{
  \item{x}{object to coerce.}

  \item{...}{further arguments passed to or from other
  methods.}
}
\value{
  If successful an object of class \code{jags_data_frame}
  otherwise NA.
}
\description{
  Coerces to an object of class \code{jags_data_frame}.
}
\seealso{
  \code{\link{jags_data_frame}} and
  \code{\link{jaggernaut}}.
}

