\name{as.jags_data_list}
\alias{as.jags_data_list}
\title{Coerce to a JAGS data list object}
\usage{
  as.jags_data_list(x, ...)
}
\arguments{
  \item{x}{object to coerce.}

  \item{...}{further arguments passed to or from other
  methods.}
}
\value{
  If successful an object of class \code{jags_data_list}
  otherwise NA.
}
\description{
  Coerces to an object of class \code{jags_data_list}.
}
\seealso{
  \code{\link{jags_data_list}} and
  \code{\link{jaggernaut}}.
}

