\name{as.jags_simulation}
\alias{as.jags_simulation}
\title{Coerce to a JAGS simulation object}
\usage{
  as.jags_simulation(x, ...)
}
\arguments{
  \item{x}{object to coerce.}

  \item{...}{further arguments passed to or from other
  methods.}
}
\value{
  If successful an object of class \code{jags_simulation}
  otherwise NA.
}
\description{
  Coerces to an object of class \code{jags_simulation}.
}
\seealso{
  \code{\link{jags_simulation}} and
  \code{\link{jaggernaut}}.
}

