% Generated by roxygen2 (4.0.2): do not edit by hand
\name{auto_corr}
\alias{auto_corr}
\title{Get auto-correlation values (s)}
\usage{
auto_corr(object, parm = "all", lags = c(1, 5, 10, 50))
}
\arguments{
\item{object}{a JAGS object}

\item{parm}{a character vector indicating the parameters for which to calculate
the auto_corr values. Either list the parmeters or use all, fixed or random.}

\item{lags}{a vector of lags at which to calculate the autocorrelation}
}
\value{
a vector, matrix or array of autocorr values
}
\description{
Get auto-correlation value(s) for JAGS objects
}
\seealso{
\code{coda::autocorr}
}

