\name{bayesian_p_value}
\alias{bayesian_p_value}
\title{Bayesian p-value}
\usage{
bayesian_p_value(x)
}
\arguments{
  \item{x}{numeric vector of posterior samples}
}
\value{
Bayesian equivalent of p-value.
}
\description{
Calculates Bayesian equivalent of frequentist p-value from
vector of posterior samples.
}

