\docType{data}
\name{burnet}
\alias{burnet}
\title{Six-spot burnet moth counts}
\format{A data frame with 665 rows and 4 columns}
\description{
  The six-spot burnet moth (\emph{Zygaena filipendulae})
  replicate counts from the Swiss butterfly monitoring
  program from Kery & Schaub (2011 p.445-446).
}
\details{
  The variables are as follows: \itemize{ \item \code{site}
  the site.  \item \code{day} the day ("season").  \item
  \code{count1} the first count.  \item \code{count2} the
  second count. }
}
\references{
  Kery M & Schaub M (2011) Bayesian Population Analysis
  using WinBUGS. Academic Press.
  (\url{http://www.vogelwarte.ch/bpa})
}
\seealso{
  \code{\link{jaggernaut}}
}
\keyword{datasets}

