% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/combine.R
\name{combine.jags_model}
\alias{combine.jags_model}
\title{Add JAGS model objects}
\usage{
\method{combine}{jags_model}(object, ...)
}
\arguments{
\item{object}{a \code{jags_model} object.}

\item{...}{additional \code{jags_model} objects to add to \code{object}.}
}
\value{
a \code{jags_model} object with multiple models
}
\description{
Adds two or more \code{jags_model} objects.
}
\seealso{
\code{\link{combine}}, \code{\link{jags_model}}
and \code{\link{jaggernaut}}.
}

