\name{combine.list}
\alias{combine.list}
\title{Combines elements of list}
\usage{
\method{combine}{list}(object, ...)
}
\arguments{
  \item{object}{a list of objects of the same class to
  combine.}

  \item{...}{additional arguments to pass to the specific
  combine function.}
}
\value{
The result of calling specific combine function.
}
\description{
Combines elements of a list by calling the specific combine
function.
}
\seealso{
\code{\link{combine}}
}

