\name{combine.mcarray}
\alias{combine.mcarray}
\title{Combines mcarrays}
\usage{
\method{combine}{mcarray}(object, ..., by = "sims")
}
\arguments{
  \item{object}{a list of objects of the same class to
  combine.}

  \item{...}{additional arguments to pass to the specific
  combine function.}

  \item{by}{string}
}
\value{
The result of calling specific combine function.
}
\description{
Combines mcarrays
}
\seealso{
\code{\link{combine}}
}

