% Generated by roxygen2 (4.0.1): do not edit by hand
\name{combine}
\alias{combine}
\title{Combines objects}
\usage{
combine(object, ...)
}
\arguments{
\item{object}{a JAGS object.}

\item{...}{additional JAGS objects to add to object.}
}
\value{
a JAGS object of the original class
}
\description{
Adds two or more JAGS object of the same class.
}

