% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/combine.R
\name{combine_list}
\alias{combine_list}
\title{Combines elements of list}
\usage{
combine_list(object, ...)
}
\arguments{
\item{object}{a list of objects of the same class to combine.}

\item{...}{additional arguments to pass to the specific combine function.}
}
\value{
The result of calling specific combine function.
}
\description{
Combines elements of a list by calling the specific combine function.
}
\seealso{
\code{\link{combine}}
}

