\name{convergence}
\alias{convergence}
\title{Get convergence value(s)}
\usage{
  convergence(object, ...)
}
\arguments{
  \item{object}{a JAGS object}

  \item{...}{passed to and from other functions}
}
\value{
  a vector, matrix or array of convergence values
}
\description{
  Get convergence values for JAGS objects
}
\seealso{
  \code{\link{convergence.jags_analysis}} and
  \code{\link{convergence.jags_power_analysis}}
}

