% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convergence-analyses.R
\name{convergence_analyses}
\alias{convergence_analyses}
\title{Combine Models}
\usage{
convergence_analyses(object, object2, parm = "all", combine = TRUE)
}
\arguments{
\item{object}{a JAGS object}

\item{object2}{A JAGS object.}

\item{parm}{a character vector indicating the parameters for which to calculate
the R-hat values. Either list the parmeters or use all, fixed or random.}

\item{combine}{a logical element indicating whether or not to summarise by 
the maximum R-hat value.}
}
\description{
Combine Models
}

