% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion-terms.R
\name{conversion_terms}
\alias{conversion_terms}
\title{Conversion Terms}
\usage{
conversion_terms(data, centre = FALSE, standardise = FALSE, dat = NULL)
}
\arguments{
\item{data}{the data frame or list of data}

\item{centre}{a logical scalar or a character vector of the variables 
to get the mean for}

\item{standardise}{a logical scalar or a character vector of the variables
to get the mean and standard deviation for}

\item{dat}{the data frame or data list to convert. If dat is NULL the dataset 
passed as the data argument is converted}
}
\value{
A matrix of the means and standard deviations for 
the centred and standardised variables
}
\description{
Get the mean and standard deviation for centred and standardised
variables.
}
\examples{
data <- data.frame(numeric = 1:10 + 0.1, integer = 1:10, 
factor = factor(1:10), date = as.Date("2000-01-01") + 1:10,
posixt = ISOdate(2000,1,1) + 1:10)
conversion_terms(data)
conversion_terms(data, centre = TRUE)
conversion_terms(data, standardise = TRUE)
conversion_terms(data, standardise = "date")
conversion_terms(data, centre = "date", standardise = "date")
}
\seealso{
\code{\link{convert_data}}
}
