% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross-cor.R
\name{cross_corr}
\alias{cross_corr}
\title{Get cross-correlation values}
\usage{
cross_corr(object, parm = "all", ...)
}
\arguments{
\item{object}{a JAGS object}

\item{parm}{a character vector indicating the parameters for which to calculate
the cross_corr values. Either list the parmeters or use all, fixed or random.}

\item{...}{passed to and from other functions}
}
\value{
a vector, matrix or array of autocorr values
}
\description{
Get cross-correlation value for JAGS objects
}
\seealso{
\code{coda::crosscorr}
}
