\name{data_jags.jags_sample}
\alias{data_jags.jags_sample}
\title{Get dataset}
\usage{
  \method{data_jags}{jags_sample} (object, ...)
}
\arguments{
  \item{object}{a \code{jags_sample} object.}

  \item{...}{further arguments passed to or from other
  methods.}
}
\value{
  The dataset.
}
\description{
  Gets the dataset from a \code{jags_sample} object.
}

