\name{data_jags}
\alias{data_jags}
\title{Get dataset(s) from a JAGS object}
\usage{
  data_jags(object, ...)
}
\arguments{
  \item{object}{a object.}

  \item{...}{other arguments.}
}
\value{
  a data.frame or list(s) of the data
}
\description{
  Gets the dataset(s) from a JAGS object.
}
\seealso{
  \code{\link{data_jags.jags_data_model}},
  \code{\link{data_jags.jags_analysis}},
  \code{\link{data_jags.jags_simulation}} and
  \code{\link{data_jags.jags_power_analysis}}
}

