% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.R
\name{date2integer}
\alias{date2integer}
\title{Date To Integer}
\usage{
date2integer(x)
}
\arguments{
\item{x}{a Date.}
}
\value{
An integer.
}
\description{
Converts Dates to integers where 
the 1st of Jan 2000 is equal to 1.
}
\examples{
date2integer(integer2date(-1:3))
}
