\docType{package}
\name{ddata}
\alias{ddata}
\alias{ddata-package}
\alias{package-ddata}
\title{Convert data sets by numericising, centring and standardising and generate dummy data.}
\description{
  ddata provides functions that make it straightforward to
  convert data into a form suitable for analysis by JAGS,
  WinBUGS or OpenBUGS and to generate dummy data sets for
  the purpose of plotting the results of statistical
  analyses.
}
\details{
  The function \code{\link{convert_data}} handles the
  numericisation, centering, and standardisation while
  \code{\link{generate_data}} generates dummy data. More
  information about each function can be found in its help
  documentation.
}

