% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddply-jags-sample.R
\name{ddply_jags_sample}
\alias{ddply_jags_sample}
\title{ddply JAGS sample}
\usage{
ddply_jags_sample(object, by, fun = sum)
}
\arguments{
\item{object}{a jags_sample object.}

\item{by}{the variables to combine by (using ddply).}

\item{fun}{the function to using when combining samples (by default fun = sum).}
}
\value{
a jags_sample object
}
\description{
Combines JAGS samples in a jags_sample object by by using function fun
}
\seealso{
\code{\link{predict.jags_analysis}}
}

