% Generated by roxygen2 (4.0.2): do not edit by hand
\name{default_model_id<-}
\alias{default_model_id<-}
\title{Set model name(s)}
\usage{
default_model_id(object) <- value
}
\arguments{
\item{object}{a JAGS object.}

\item{value}{a string or number defining the default model id to use in predictions etc
of the BUGS language.}
}
\description{
Set the default_model_id component of a JAGS object.
}
\seealso{
\code{\link{jaggernaut}}
}

