% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derived-code.R
\name{derived_code<-}
\alias{derived_code<-}
\title{Set derived code in a JAGS object}
\usage{
derived_code(object) <- value
}
\arguments{
\item{object}{a JAGS object.}

\item{value}{a character element or NULL.}
}
\value{
The replacement method changes the derived_code component of the object.
}
\description{
Sets the derived_code component of a JAGS object.
}
\seealso{
\code{\link{derived_code}} and \code{\link{jaggernaut}}
}
