% Generated by roxygen2 (4.0.2): do not edit by hand
\name{dic_jags.jags_analysis}
\alias{dic_jags.jags_analysis}
\title{DIC values}
\usage{
\method{dic_jags}{jags_analysis}(object, ...)
}
\arguments{
\item{object}{a \code{jags_analysis}.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
The DIC values as a data.frame.
}
\description{
Gets Deviance Information Criterion (DIC) values from a JAGS analysis.
}
\seealso{
\code{\link{jaggernaut}}
}

