\name{dic_jags}
\alias{dic_jags}
\title{DIC values}
\usage{
  dic_jags(object, ...)
}
\arguments{
  \item{object}{a JAGS object.}

  \item{...}{further arguments passed to or from other
  methods.}
}
\value{
  The DIC values as a data.frame.
}
\description{
  Gets Deviance Information Criterion values from a JAGS
  object.
}
\seealso{
  \code{\link{jaggernaut}}
}

