\name{extract_data<-}
\alias{extract_data<-}
\title{Set extract data}
\usage{
  extract_data(object) <- value
}
\arguments{
  \item{object}{a JAGS object.}

  \item{value}{a function or NULL.}
}
\value{
  The replacement method changes the extract_data component
  of the object.
}
\description{
  Sets the extract_data component of a JAGS object.
}
\seealso{
  \code{\link{extract_data}} and \code{\link{jaggernaut}}
}

