\name{extract_data}
\alias{extract_data}
\title{Get extract data}
\usage{
extract_data(object, ...)
}
\arguments{
  \item{object}{a JAGS object.}

  \item{...}{further arguments passed to or from other
  methods.}
}
\value{
The extract_data component of a JAGS object which is either
a function or NULL.
}
\description{
Get the extract_data component of a JAGS object.
}
\seealso{
\code{\link{jaggernaut}}
}

