% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-falcon.R
\docType{data}
\name{falcon}
\alias{falcon}
\title{Falcon breeding population data}
\format{A data frame with 40 rows and 4 columns
\describe{
  \item{Year}{the year (integer).}
  \item{Pairs}{the number of adult pairs (integer).}
  \item{R.pairs}{the number of reproductive pairs (integer).}
  \item{Eyasses}{the number of fledged young (integer).}
}}
\usage{
falcon
}
\description{
Simulated falcon population
breeding data.
}
\keyword{datasets}

