% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-fritillary.R
\docType{data}
\name{fritillary}
\alias{fritillary}
\title{Fritillary butterfly abundance data}
\format{A data frame with 665 rows and 4 columns}
\usage{
fritillary
}
\description{
The silver-washed fritillary (\emph{Argynnis paphia}) 
butterfly duplicate site counts from 
Kery & Schaub (2011 p.396).
}
\details{
The variables are as follows:
\itemize{
  \item \code{site} the site surveyed.
  \item \code{day} the day of the survey.
  \item \code{count1} the first count.
  \item \code{count2} the second count.
}
}
\references{
Kery M & Schaub M (2011) Bayesian Population Analysis
using WinBUGS. Academic Press. (\url{http://www.vogelwarte.ch/bpa})
}
\seealso{
\code{\link{jaggernaut}}
}
\keyword{datasets}

