\name{gen_inits<-}
\alias{gen_inits<-}
\title{Set generate initial values}
\usage{
gen_inits(object) <- value
}
\arguments{
  \item{object}{a JAGS object.}

  \item{value}{a function or NULL.}
}
\description{
Sets the gen_inits component of a JAGS object.
}
\seealso{
\code{\link{gen_inits}} and \code{\link{jaggernaut}}
}

