\name{ind}
\alias{ind}
\title{Get R-hat value(s)}
\usage{
  ind(object, ...)
}
\arguments{
  \item{object}{a JAGS object}

  \item{...}{passed to and from other functions}
}
\value{
  a vector, matrix or array of ind values
}
\description{
  Get R-hat value or values for JAGS objects
}
\seealso{
  \code{\link{ind.jags_analysis}} and
  \code{\link{ind.jags_power_analysis}}
}

