% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.R
\name{integer2date}
\alias{integer2date}
\title{Integer To Date}
\usage{
integer2date(x)
}
\arguments{
\item{x}{an integer}
}
\value{
A Date.
}
\description{
Converts integers to Dates where 
the 1st of Jan 2000 is equal to 1.
}
\examples{
integer2date(-1:3)
}

