% Generated by roxygen2 (4.0.1): do not edit by hand
\name{is.jags_aggregation}
\alias{is.jags_aggregation}
\title{Test for objects of class jags_aggregation}
\usage{
is.jags_aggregation(x)
}
\arguments{
\item{x}{the object to be tested.}
}
\value{
a logical element indicating whether or not \code{x} is of
class \code{jags_aggregation}.
}
\description{
Tests for objects of class \code{jags_aggregation}.
}

