\name{is.jags_data}
\alias{is.jags_data}
\title{Test for objects of class jags_data}
\usage{
  is.jags_data(x)
}
\arguments{
  \item{x}{the object to be tested.}
}
\value{
  a logical element indicating whether or not \code{x} is
  of class \code{jags_data}.
}
\description{
  Tests for objects of class \code{jags_data}.
}

