\name{is.jags_data_list}
\alias{is.jags_data_list}
\title{Test for objects of class jags_data_list}
\usage{
  is.jags_data_list(x)
}
\arguments{
  \item{x}{the object to be tested.}
}
\value{
  a logical element indicating whether or not \code{x} is
  of class \code{jags_data_list}.
}
\description{
  Tests for objects of class \code{jags_data_list}.
}
\seealso{
  \code{\link{as.jags_data_list}} and
  \code{\link{jaggernaut}}.
}

