\name{is.jags_power_analysis}
\alias{is.jags_power_analysis}
\title{Test for objects of class jags_power_analysis}
\usage{
  is.jags_power_analysis(x)
}
\arguments{
  \item{x}{the object to be tested.}
}
\value{
  a logical element indicating whether or not \code{x} is
  of class \code{jags_power_analysis}.
}
\description{
  Tests for objects of class \code{jags_power_analysis}.
}
\seealso{
  \code{\link{jags_power_analysis}} and
  \code{\link{jaggernaut}}.
}

