\name{is.jags_simulation}
\alias{is.jags_simulation}
\title{Test for objects of class jags_simulation}
\usage{
is.jags_simulation(x)
}
\arguments{
  \item{x}{the object to be tested.}
}
\value{
a logical element indicating whether or not \code{x} is of
class \code{jags_simulation}.
}
\description{
Tests for objects of class \code{jags_simulation}.
}
\seealso{
\code{\link{jags_simulation}} and \code{\link{jaggernaut}}.
}

