% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is_converted_data}
\alias{is_converted_data}
\title{Test data set}
\usage{
is_converted_data(x)
}
\arguments{
\item{x}{the object to test}
}
\value{
A flag (logical scalar) indicating whether converted data
}
\description{
Tests whether x is converted data that can be input into
JAGS, WinBUGS or OpenBUGS.
}
\examples{
is_converted_data(trees)
is_converted_data(convert_data(trees))
is_converted_data(list(x = 1:10, y = 0.33, z = matrix(1:9, ncol = 3)))
is_converted_data(list(x = 1:10, y = 0.33, z = matrix(1:9, ncol = 3)))
is_converted_data(list(x = factor(1:10), y = 0.33, z = matrix(1:9, ncol = 3)))
}
\seealso{
\code{\link{is_convertible_data}} and 
\code{\link{convert_data}}
}

