% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is_convertible_data}
\alias{is_convertible_data}
\title{Test data set}
\usage{
is_convertible_data(x)
}
\arguments{
\item{x}{the object to test}
}
\value{
A logical scalar
}
\description{
Tests whether x is a data.frame or list of data suitable for conversion for 
input into 
JAGS, WinBUGS or OpenBUGS.
}
\examples{
is_convertible_data(trees)
is_convertible_data(as.list(trees))
trees$Comment <- "text"
is_convertible_data(trees)
}
\seealso{
\code{\link{is_convertible_data_list}} and 
\code{\link{is_convertible_data_frame}}
}

