\docType{package}
\name{jaggernaut}
\alias{jaggernaut}
\alias{jaggernaut-package}
\alias{package-jaggernaut}
\title{Bayesian analysis with JAGS}
\description{
  The package jaggernaut provides functions that facilitate
  Bayesian analyses using JAGS (Just Another Gibbs
  Sampler).
}
\details{
  For additional information see \code{\link{model}} and
  \code{\link{analysis}}.

  For demonstrative purposes the real data sets in Kery and
  Schaub's excellent book Bayesian Population Analysis
  using WinBUGS: A hierarchical perspective have been
  imported into the package and their example analyses are
  in the process of being implemented. To date analyses
  have been implemented on the \code{\link{peregrine}},
  \code{\link{tits}} and \code{\link{hm}} data sets.

  This is a package in beta form. Please pass on any
  comments to the package maintainer.
}
\references{
  Kery M & Schaub M (2011) Bayesian Population Analysis
  using WinBUGS. Academic Press.
  (\url{http://www.vogelwarte.ch/bpa})

  Plummer M (2012) JAGS Version 3.3.0 User Manual
  \url{http://sourceforge.net/projects/mcmc-jags/files/Manuals/}

  Thorley JL (2013) jaggernaut: An R package to facilitate
  Bayesian analyses using JAGS (Just Another Gibbs Sampler)
}
\seealso{
  \code{\link{model}}, \code{\link{analysis}},
  \code{\link{convergence}}, \code{\link{estimates}},
  \code{\link{derived}}, \code{\link{plot.janalysis}},
  \code{\link{bluebug}}, \code{\link{burnet}},
  \code{\link{fritillary}}, \code{\link{hm}},
  \code{\link{leisleri}}, \code{\link{orchids}},
  \code{\link{owls}}, \code{\link{p610}},
  \code{\link{peregrine}}, \code{\link{pinna}},
  \code{\link{tits}}
}

