\name{jags_data_frame}
\alias{jags_data_frame}
\title{JAGS data frame}
\usage{
  jags_data_frame(x)
}
\arguments{
  \item{x}{a data.frame from which to generate a
  \code{jags_data_frame} object.}
}
\value{
  A \code{jags_data_frame} object.
}
\description{
  Creates an object of class \code{jags_data_frame}.
}
\seealso{
  \code{\link{jags_data}}, \code{\link{jags_data_list}} and
  \code{\link{jaggernaut}}.
}

