\name{jags_data_list}
\alias{jags_data_list}
\title{JAGS data list}
\usage{
  jags_data_list(x)
}
\arguments{
  \item{x}{a named list.}
}
\value{
  A \code{jags_data_list} object.
}
\description{
  Creates an object of class \code{jags_data_list}.
}
\seealso{
  \code{\link{jags_data}}, \code{\link{jags_data_frame}}
  and \code{\link{jaggernaut}}.
}

