\name{jags_sample}
\alias{jags_sample}
\title{JAGS sample}
\usage{
  jags_sample(chains, parm, data)
}
\arguments{
  \item{chains}{a \code{jagr_sample} object.}

  \item{parm}{a character scalar.}

  \item{data}{a data.frame.}
}
\value{
  A jags_sample object.
}
\description{
  An object of class \code{jags_sample}.
}
\seealso{
  \code{\link{predict.jags_analysis}} and
  \code{\link{jaggernaut}}.
}

