% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blocks.R
\name{jg_block_names<-}
\alias{jg_block_names<-}
\title{Set block names in JAGS model code}
\usage{
jg_block_names(x) <- value
}
\arguments{
\item{x}{string of JAGS model code}

\item{value}{character vector of block names}
}
\value{
Modified JAGS model code.
}
\description{
Sets block names in JAGS model code.
Throws an error if unbalanced brackets.
Strips out comments.
}
\examples{
 x <- "data {X <- 2} model { Y ~ dpois(X) }"
jg_block_names(x) <- c("settings", "model")
print(x)
}

