% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blocks.R
\name{jg_block_names}
\alias{jg_block_names}
\title{Get block names in JAGS model code}
\usage{
jg_block_names(x)
}
\arguments{
\item{x}{string of JAGS model code}
}
\value{
A character vector of block names in order occur in code.
}
\description{
Get block names in JAGS model code
}
\examples{
jg_block_names("data {X <- 2} model { Y ~ dpois(X) }")
}

