% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blocks.R
\name{jg_blocks}
\alias{jg_blocks}
\title{Parses blocks in JAGS model code}
\usage{
jg_blocks(x)
}
\arguments{
\item{x}{string of JAGS model code}
}
\value{
A named character vector where the names are the block names.
Throws an error if the brackets are unbalanced.
}
\description{
Converts blocks in JAGS model code into a named
character vector where the names are the block names.
}
\examples{
jg_blocks("data {X <- 2} model { Y ~ dpois(X) }")
}

