% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{jg_dists}
\alias{jg_dists}
\title{Get distributions in JAGS model code}
\usage{
jg_dists(x)
}
\arguments{
\item{x}{string of JAGS model code}
}
\value{
Sorted character vector of distributions
that occur one or more times.
}
\description{
Gets distributions in JAGS model code.
A distribution is any JAGS word between ~ and (.
Where a JAGS word must start with a letter
and can only contain letters, numbers, '.'
and '_'.
}

