% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix.R
\name{jg_fix}
\alias{jg_fix}
\title{Fix JAGS model code}
\usage{
jg_fix(x, extended = FALSE)
}
\arguments{
\item{x}{string of JAGS model code}

\item{extended}{flag of whether to allow extended BUGS language.}
}
\value{
If successful a string of the fixed code. 
If fails issues warning and returns FALSE.
}
\description{
Fixes JAGS model code. Currently only
drops blocks with invalid names and reorders blocks 
(JAGS requires blocks to be ordered).
}

