% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{jg_format}
\alias{jg_format}
\title{Formats JAGS model code}
\usage{
jg_format(x)
}
\arguments{
\item{x}{string of JAGS model code}
}
\value{
String of formatted model code or FALSE if fails 
(in which case also gives a warning).
}
\description{
Formats JAGS model code in the JAGS dialect of the BUGS language
(extended or string).
}

