% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{jg_funcs}
\alias{jg_funcs}
\title{Get functions in JAGS model code}
\usage{
jg_funcs(x)
}
\arguments{
\item{x}{string of JAGS model code}
}
\value{
Character vector of functions.
}
\description{
Gets functions in JAGS model code
A function is any JAGS word followed by a '('
that is not preceeded by a '~'.
Where a JAGS word must start with a letter
and can only contain letters, numbers, '.'
and '_'.
}

