% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{jg_vnodes}
\alias{jg_vnodes}
\title{Get variable nodes in JAGS model code}
\usage{
jg_vnodes(x, type = "both", indices = FALSE, comment = "[$][^\\n]+[$]")
}
\arguments{
\item{x}{string of JAGS model code}

\item{type}{string of node type. Must be 'stochastic', 'deterministic'
or 'both' (the default). A variable node is any word possibly}

\item{indices}{flag of whether to retain indices}

\item{comment}{A string defining the regular expression to use to 
filter text from the comment for the node.}
}
\value{
Character vector of unique sorted variable node names
}
\description{
Gets names of variable (as opposed to constant) named nodes in JAGS model code
}
\examples{
model_code <- "data{
   Y ~ dpois(2) 
}

model {

 bLambda ~ dlnorm(0,10^-2) #$\\\\lamda_{beta}$
 for (i in 1:length(x)) { x[i]~dpois(bLambda) 
   b[i] ~dpois(1)
   bc[i] <- b[i]
 }
 bd <- dpois(1, 1)
}"
jg_vnodes(model_code)

}

