% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data-leisleri.R
\docType{data}
\name{leisleri}
\alias{leisleri}
\title{Leisler's bats survival data}
\format{A data frame with 181 rows and 19 columns}
\usage{
leisleri
}
\description{
The female Leisler's bats (\emph{Nyctalus leisleri})
capture-recapture data collected by Wigbert Schorcht and colleagues
in Thuringia (Germany) from 1989 to 2008 from
Kery & Schaub (2011 p.231).
}
\details{
The variables are as follows:
\itemize{
  \item \code{V1, V2, ..., V19} an indicator variable specify whether or not the
  individual was 0 = undetected or 1 = detected.
}
}
\references{
Kery M & Schaub M (2011) Bayesian Population Analysis
using WinBUGS. Academic Press. (\url{http://www.vogelwarte.ch/bpa})
}
\seealso{
\code{\link{jaggernaut}}
}
\keyword{datasets}

