% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/merge-jags-samples.R
\name{merge_jags_samples}
\alias{merge_jags_samples}
\title{Merge JAGS samples}
\usage{
merge_jags_samples(object, by = NULL, fun = sum)
}
\arguments{
\item{object}{a list of jags_sample objects.}

\item{by}{the variables to combine by (using merge).}

\item{fun}{the function to using when combining samples (by default fun = sum).}
}
\value{
a jags_sample object
}
\description{
Merges JAGS samples in multiple jags_sample objects by by using function fun
}
\seealso{
\code{\link{predict.jags_analysis}} and
\code{\link{ddply_jags_sample}}
}

