% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/model.R
\name{model}
\alias{model}
\title{Get model}
\usage{
model(object, ...)
}
\arguments{
\item{object}{a JAGS object.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
The model component of a JAGS object.
}
\description{
Get the model component of a JAGS object.
}
\seealso{
\code{\link{jaggernaut}}
}

