\name{model_code}
\alias{model_code}
\title{Get model code}
\usage{
  model_code(object, ...)
}
\arguments{
  \item{object}{a JAGS object.}

  \item{...}{further arguments passed to or from other
  methods.}
}
\value{
  The model_code component of a JAGS object.
}
\description{
  Get the model_code component of a JAGS object.
}
\seealso{
  \code{\link{jaggernaut}}
}

