% Generated by roxygen2 (4.0.2): do not edit by hand
\name{model_id}
\alias{model_id}
\title{Get model id}
\usage{
model_id(object, reference = FALSE, ...)
}
\arguments{
\item{object}{a JAGS object.}

\item{reference}{flag indicating whether to return the assigned unique reference
names or the original names (the default)}

\item{...}{further arguments passed to or from other methods.}
}
\value{
The model_id component of a JAGS object.
}
\description{
Get the model_id component of a JAGS object.
}
\seealso{
\code{\link{jaggernaut}}
}

