% Generated by roxygen2 (4.0.2): do not edit by hand
\name{model_name<-}
\alias{model_name<-}
\title{Set model name(s)}
\usage{
model_name(object) <- value
}
\arguments{
\item{object}{a JAGS object.}

\item{value}{a string defining the model name in the JAGS dialect
of the BUGS language.}
}
\description{
Set the model_name component of a JAGS object.
}
\seealso{
\code{\link{jaggernaut}}
}

