\name{modify_data<-}
\alias{modify_data<-}
\title{Set modify data}
\usage{
modify_data(object) <- value
}
\arguments{
  \item{object}{a JAGS object.}

  \item{value}{a function or NULL to replace the modify
  data fnction.}
}
\description{
Set the modify_data component of a JAGS object.
}
\seealso{
\code{\link{jaggernaut}}
}

