% Generated by roxygen2 (4.0.2): do not edit by hand
\name{modify_data_aggregation<-}
\alias{modify_data_aggregation<-}
\title{Set modify data derived}
\usage{
modify_data_aggregation(object) <- value
}
\arguments{
\item{object}{a JAGS object.}

\item{value}{a function or NULL to replace the modify data derived fnction.}
}
\description{
Set the modify_data_aggregation component of a JAGS object.
}
\seealso{
\code{\link{jaggernaut}}
}

