% Generated by roxygen2 (4.0.1): do not edit by hand
\name{modify_data_aggregation}
\alias{modify_data_aggregation}
\title{Get modify data derived}
\usage{
modify_data_aggregation(object, ...)
}
\arguments{
\item{object}{a JAGS object.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
The modify_data_aggregation component of a JAGS object.
}
\description{
Get the modify_data_aggregation component of a JAGS object.
}
\seealso{
\code{\link{jaggernaut}}
}

