\name{modify_data_derived<-}
\alias{modify_data_derived<-}
\title{Set modify data derived}
\usage{
modify_data_derived(object) <- value
}
\arguments{
  \item{object}{a JAGS object.}

  \item{value}{a function or NULL to replace the modify
  data derived fnction.}
}
\description{
Set the modify_data_derived component of a JAGS object.
}
\seealso{
\code{\link{jaggernaut}}
}

