% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify-data-derived.R
\name{modify_data_derived}
\alias{modify_data_derived}
\title{Get modify data derived}
\usage{
modify_data_derived(object, ...)
}
\arguments{
\item{object}{a JAGS object.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
The modify_data_derived component of a JAGS object.
}
\description{
Get the modify_data_derived component of a JAGS object.
}
\seealso{
\code{\link{jaggernaut}}
}
