% Generated by roxygen2 (4.0.2): do not edit by hand
\name{nanalyses}
\alias{nanalyses}
\title{Number of analyses in a JAGS object}
\usage{
nanalyses(object, ...)
}
\arguments{
\item{object}{a JAGS object}

\item{...}{further arguments passed to or from other methods.}
}
\value{
an integer element indicating the number of analyses in object
}
\description{
Gets the number of analyses in a JAGS object
}

