\name{nchains.jags_analysis}
\alias{nchains.jags_analysis}
\title{Number of MCMC chains in a JAGS analysis}
\usage{
  \method{nchains}{jags_analysis} (object, ...)
}
\arguments{
  \item{object}{a jags_analysis object}

  \item{...}{other arguments passed to generic function.}
}
\description{
  Gets the number of MCMC chains in a JAGS analysis object.
}
\seealso{
  \code{\link{nchains}} and \code{\link{jags_analysis}}
}

