\name{nchains.jags_power_analysis}
\alias{nchains.jags_power_analysis}
\title{Number of MCMC chains in a JAGS power analysis}
\usage{
  \method{nchains}{jags_power_analysis} (object, ...)
}
\arguments{
  \item{object}{a jags_power_analysis object}

  \item{...}{other arguments passed to generic function.}
}
\description{
  Gets the number of MCMC chains in a JAGS power analysis
  object.
}
\seealso{
  \code{\link{nchains}} and
  \code{\link{jags_power_analysis}}
}

