% Generated by roxygen2 (4.0.1): do not edit by hand
\name{nchains}
\alias{nchains}
\title{Number of MCMC chains in a JAGS object}
\usage{
nchains(object)
}
\arguments{
\item{object}{a JAGS object}
}
\value{
an integer element indicating the number of MCMC chains in object
}
\description{
Gets the number of MCMC chains in a JAGS object
}

