% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/niters.R
\name{niters}
\alias{niters}
\title{Number of MCMC iterations used to generate a JAGS object}
\usage{
niters(object)
}
\arguments{
\item{object}{a JAGS object}
}
\value{
an integer element indicating the number of MCMC iterations used to generate object
}
\description{
Gets the number of MCMC iterations used to generate a JAGS object
}

