\name{nmodel.jags_analysis}
\alias{nmodel.jags_analysis}
\title{Number of models in a JAGS analysis}
\usage{
  \method{nmodel}{jags_analysis} (object, ...)
}
\arguments{
  \item{object}{a jags_analysis object}

  \item{...}{other arguments passed to generic function.}
}
\description{
  Gets the number of models in a JAGS analysis object.
}
\seealso{
  \code{\link{nmodel}} and \code{\link{jags_analysis}}
}

