% Generated by roxygen2 (4.0.1): do not edit by hand
\name{nreps}
\alias{nreps}
\title{Number of replicates in a JAGS object}
\usage{
nreps(object)
}
\arguments{
\item{object}{a JAGS object}
}
\value{
an integer element indicating the number of replicates in object
}
\description{
Gets the number of replicates in a JAGS object
}

