% Generated by roxygen2 (4.0.2): do not edit by hand
\name{nsims}
\alias{nsims}
\title{Number of MCMC simulations in a JAGS object}
\usage{
nsims(object)
}
\arguments{
\item{object}{a JAGS object}
}
\value{
an integer element indicating the number of MCMC simulations in object
}
\description{
Gets the number of MCMC simulations in a JAGS object
}

