\name{number_of_chains.jags_power_analysis}
\alias{number_of_chains.jags_power_analysis}
\title{Number of MCMC chains in a JAGS power analysis}
\usage{
  \method{number_of_chains}{jags_power_analysis} (object,
    ...)
}
\arguments{
  \item{object}{a jags_power_analysis object}

  \item{...}{other arguments passed to generic function.}
}
\description{
  Gets the number of MCMC chains in a JAGS power analysis
  object.
}
\seealso{
  \code{\link{number_of_chains}} and
  \code{\link{jags_power_analysis}}
}

