\name{number_of_chains}
\alias{nchains}
\alias{number_of_chains}
\title{Number of MCMC chains in a JAGS object}
\usage{
  number_of_chains(object, ...)
}
\arguments{
  \item{object}{a JAGS object}
}
\value{
  an integer element indicating the number of MCMC chains
  in object
}
\description{
  Gets the number of MCMC chains in a JAGS object
}

