\name{number_of_iterations.jags_analysis}
\alias{number_of_iterations.jags_analysis}
\title{Number of iterations in a JAGS analysis}
\usage{
  \method{number_of_iterations}{jags_analysis} (object,
    ...)
}
\arguments{
  \item{object}{a jags_analysis object}

  \item{...}{other arguments passed to generic function.}
}
\description{
  Gets the number of iterations in a JAGS analysis object.
}
\seealso{
  \code{\link{number_of_iterations}} and
  \code{\link{jags_analysis}}
}

