\name{number_of_iterations}
\alias{niter}
\alias{number_of_iterations}
\title{Number of MCMC iterations in a JAGS object}
\usage{
  number_of_iterations(object, ...)
}
\arguments{
  \item{object}{a JAGS object}
}
\value{
  an integer element indicating the number of MCMC
  iterations in object
}
\description{
  Gets the number of MCMC iterations in a JAGS object
}
\seealso{
  \code{\link{number_of_iteration.jags_analysis}} and
  \code{\link{number_of_iterations.jags_power_analysis}}
}

