\name{number_of_models}
\alias{nmodel}
\alias{number_of_models}
\title{Number of models in a JAGS object}
\usage{
  number_of_models(object, ...)
}
\arguments{
  \item{object}{a JAGS object}
}
\value{
  an integer element indicating the number of models in
  object
}
\description{
  Gets the number of models in a JAGS object
}

