\name{number_of_replicates.jags_simulation}
\alias{number_of_replicates.jags_simulation}
\title{Number of replicates in a JAGS simulation}
\usage{
  \method{number_of_replicates}{jags_simulation} (object,
    ...)
}
\arguments{
  \item{object}{a jags_simulation object}

  \item{...}{other arguments passed to generic function.}
}
\description{
  Gets the number of replicates in a JAGS simulation object
}
\seealso{
  \code{\link{number_of_replicates}} and
  \code{\link{jags_simulation}}
}

