\name{number_of_replicates}
\alias{nrep}
\alias{number_of_replicates}
\title{Number of replicates in a JAGS object}
\usage{
  number_of_replicates(object, ...)
}
\arguments{
  \item{object}{a JAGS object}
}
\value{
  an integer element indicating the number of replicates in
  object
}
\description{
  Gets the number of replicates in a JAGS object
}
\seealso{
  \code{\link{number_of_replicates.jags_simulation}} and
  \code{\link{number_of_replicates.jags_power}}
}

