\name{number_of_simulations}
\alias{nsim}
\alias{number_of_simulations}
\title{Number of MCMC simulations in a JAGS object}
\usage{
  number_of_simulations(object, ...)
}
\arguments{
  \item{object}{a JAGS object}
}
\value{
  an integer element indicating the number of MCMC
  simulations in object
}
\description{
  Gets the number of MCMC simulations in a JAGS object
}
\seealso{
  \code{\link{number_of_simulations.jags_analysis}} and
  \code{\link{number_of_simulations.jags_power_analysis}}
}

