\name{number_of_values.jags_simulation}
\alias{number_of_values.jags_simulation}
\title{Number of values in a JAGS simulation}
\usage{
  \method{number_of_values}{jags_simulation} (object, ...)
}
\arguments{
  \item{object}{a jags_simulation object}

  \item{...}{other arguments passed to generic function.}
}
\description{
  Gets the number of values in a JAGS simulation object
}
\seealso{
  \code{\link{number_of_values}} and
  \code{\link{jags_simulation}}
}

