\name{number_of_values}
\alias{number_of_values}
\alias{nvalue}
\title{Number of simulation values in a JAGS object}
\usage{
  number_of_values(object, ...)
}
\arguments{
  \item{object}{a JAGS object}
}
\value{
  an integer element indicating the number of simulation
  values in object
}
\description{
  Gets the number of simulation values in a JAGS object
}
\seealso{
  \code{\link{number_of_values.jags_simulation}} and
  \code{\link{number_of_values.jags_power}}
}

