\name{nvalues}
\alias{nvalues}
\title{Number of simulation values in a JAGS object}
\usage{
nvalues(object)
}
\arguments{
  \item{object}{a JAGS object}
}
\value{
an integer element indicating the number of simulation
values in object
}
\description{
Gets the number of simulation values in a JAGS object
}

