\docType{data}
\name{orchids}
\alias{orchids}
\title{Showy lady's slipper capture-recapture data}
\format{A data frame with 250 rows and 11 columns}
\description{
  The showy lady's slipper (\emph{Cypripedium reginae})
  capture-recapture data collected by Kathy Gregg in Big
  Draft (West Virginia) from 1989 to 1999 from Kery &
  Schaub (2011 p.166).
}
\details{
  The variables are as follows: \itemize{ \item \code{V1,
  V2, ..., V11} the state of the individual plant by year
  where 0 = undetected, 1 = vegetative and 2 = flowering. }
}
\references{
  Kery M & Schaub M (2011) Bayesian Population Analysis
  using WinBUGS. Academic Press.
  (\url{http://www.vogelwarte.ch/bpa})
}
\keyword{datasets}

