% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{owls}
\alias{owls}
\title{Long-eared owl detections}
\format{A data frame with 40 rows and 11 columns}
\usage{
owls
}
\description{
The long-eared owl (\emph{Asio otus})
territory replicate detections from the 2009 breeding season
by Simon Birrer from Kery & Schaub (2011 p.454).
}
\details{
The variables are as follows:
\itemize{
  \item \code{territory} the territory.
  \item \code{obs1, obs2, ..., obs5} the count by visit.
  \item \code{date1, date2, ..., date5} the day of the year by visit.
}
}
\references{
Kery M & Schaub M (2011) Bayesian Population Analysis
using WinBUGS. Academic Press. (\url{http://www.vogelwarte.ch/bpa})
}
\seealso{
\code{\link{jaggernaut}}
}
\keyword{datasets}

