% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string.R
\name{paste_names}
\alias{paste_names}
\title{Paste Names}
\usage{
paste_names(x, quotes = FALSE)
}
\arguments{
\item{x}{character vector}

\item{quotes}{flag indicating whether to quote names}
}
\value{
String of pasted names
}
\description{
Paste names
}
\examples{
paste_names(c("x","y","z"))
paste_names(c("x","y","z"), TRUE)
}

