% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-peregrine.R
\docType{data}
\name{peregrine}
\alias{peregrine}
\title{Peregrine falcon breeding population data}
\format{A data frame with 40 rows and 4 columns}
\usage{
peregrine
}
\description{
Simulated peregrine falcon population
breeding from 1964 to 2003.

 The variables are as follows:
\itemize{
  \item \code{Year} the year (integer).
  \item \code{Pairs} the number of adult pairs (integer).
  \item \code{R.pairs} the number of reproductive pairs (integer).
  \item \code{Eyasses} the number of fledged young (integer).
}
}
\examples{
data(peregrine)
summary(peregrine)
}
\keyword{datasets}

