% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-pinna.R
\docType{data}
\name{pinna}
\alias{pinna}
\title{Pen shell detection data}
\format{A data frame with 143 rows and 3 columns}
\source{
Kery & Schaub (2011 p.166) courtesy of Iris Hendriks and colleagues
}
\usage{
pinna
}
\description{
The pen shell (\emph{Pinna nobilis}) detection data from the Balearic Islands 
in 2010.

The variables are as follows:
\itemize{
  \item \code{d1} indicator for shell detected by first team.
  \item \code{d2} indicator for shell detected by second team.
  \item \code{width} shell width (cm).
}
}
\examples{
data(pinna)
summary(pinna)
# demo(pina)
}
\references{
Kery M & Schaub M (2011) Bayesian Population Analysis
using WinBUGS. Academic Press. (\url{http://www.vogelwarte.ch/bpa})
}
\keyword{datasets}

