% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot.jags_discrepancies}
\alias{plot.jags_discrepancies}
\title{Plot a JAGS discrepancies object}
\usage{
\method{plot}{jags_discrepancies}(x, ...)
}
\arguments{
\item{x}{a jags_discrepancies object}

\item{...}{additional arguments}
}
\description{
Plots a JAGS discrepancies object
}

