% Generated by roxygen2 (4.0.1): do not edit by hand
\name{power_jags}
\alias{power_jags}
\title{JAGS power}
\usage{
power_jags(object, parm = c(fixed = 0), level = "current",
  estimate = "current")
}
\arguments{
\item{object}{a \code{jags_power_analysis} object.
for which to calculate the power.}

\item{parm}{a named list of the parameters and the value
for which to calculate the power.}

\item{level}{a numeric scalar specifying the significance level or a character
scalar specifying which mode the level should be taken from. By default the
level is as currently specified by \code{opts_jagr} in the global options.}

\item{estimate}{a character scalar specifying whether the point estimate should
be the "mean" or the "median" or a character scalar which mode the level should be #' taken from. By default the
estimate is as currently specified by \code{opts_jagr} in the global options.}
}
\value{
A \code{data.frame} of the power estimates.
}
\description{
Extracts the power estimates from a \code{jags_power_analysis} object
}
\seealso{
\code{\link{jags_power_analysis}}
and \code{\link{jaggernaut}}
}

