\name{print.summary_janalysis}
\alias{print.summary_janalysis}
\title{Print summary of JAGS analysis (janalysis) object}
\usage{
  \method{print}{summary_janalysis} (x, ...)
}
\arguments{
  \item{x}{the summary of a JAGS analysis (janalysis)
  object.}

  \item{...}{further arguments pass to or from other
  methods.}
}
\description{
  Prints a summary of a JAGS analysis (janalysis) object
}
\seealso{
  \code{\link{jaggernaut}}
}

