% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random-effects.R
\name{random_effects<-}
\alias{random_effects<-}
\title{Set random effects}
\usage{
random_effects(object) <- value
}
\arguments{
\item{object}{a JAGS object.}

\item{value}{a named list vector of the random effects or NULL.}
}
\description{
Set the random_effects component of a JAGS object.
}
\seealso{
\code{\link{jaggernaut}}
}

