% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string.R
\name{reverse_strings}
\alias{reverse_strings}
\title{Reverse Strings}
\usage{
reverse_strings(x)
}
\arguments{
\item{x}{character vector}
}
\value{
Reversed character vector
}
\description{
Reverses strings
}
\examples{
reverse_strings(c("abc","def"))
}
