\name{rhat}
\alias{rhat}
\title{Get R-hat value(s)}
\usage{
  rhat(object, ...)
}
\arguments{
  \item{object}{a JAGS object}

  \item{...}{passed to and from other functions}
}
\value{
  a vector or matrix of rhat values
}
\description{
  Get R-hat value or values for JAGS objects
}
\seealso{
  \code{\link{rhat.jags_analysis}} and
  \code{\link{rhat.jags_power_analysis}}
}

