\name{select<-}
\alias{select<-}
\title{Set select}
\usage{
select(object) <- value
}
\arguments{
  \item{object}{a JAGS object.}

  \item{value}{a character vector of the variables to
  select or NULL.}
}
\description{
Set the select component of a JAGS object.
}
\seealso{
\code{\link{jaggernaut}}
}

