% Generated by roxygen2 (4.0.1): do not edit by hand
\name{select_aggregation}
\alias{select_aggregation}
\title{Get select derived}
\usage{
select_aggregation(object, ...)
}
\arguments{
\item{object}{a JAGS object.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
The select_aggregation component of a JAGS object.
}
\description{
Get the select_aggregation component of a JAGS object.
}
\seealso{
\code{\link{jaggernaut}}
}

