% Generated by roxygen2 (4.0.2): do not edit by hand
\name{select_data}
\alias{select_data}
\title{Get select data}
\usage{
select_data(object, ...)
}
\arguments{
\item{object}{a JAGS object.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
The select_data component of a JAGS object.
}
\description{
Get the select_data component of a JAGS object.
}
\seealso{
\code{\link{jaggernaut}}
}

