% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select-data-derived.R
\name{select_data_derived<-}
\alias{select_data_derived<-}
\title{Set modify data derived}
\usage{
select_data_derived(object) <- value
}
\arguments{
\item{object}{a JAGS object.}

\item{value}{a function or NULL to replace the modify data derived fnction.}
}
\description{
Set the select_data_derived component of a JAGS object.
}
\seealso{
\code{\link{jaggernaut}}
}

