% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select-data-derived.R
\name{select_data_derived}
\alias{select_data_derived}
\title{Get select derived}
\usage{
select_data_derived(object, ...)
}
\arguments{
\item{object}{a JAGS object.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
The select_data_derived component of a JAGS object.
}
\description{
Get the select_data_derived component of a JAGS object.
}
\seealso{
\code{\link{jaggernaut}}
}
