\name{values}
\alias{values}
\title{Get values}
\usage{
values(object, ...)
}
\arguments{
  \item{object}{a JAGS object.}

  \item{...}{further arguments passed to or from other
  methods.}
}
\value{
The values component of a JAGS object.
}
\description{
Get the values component of a JAGS object.
}
\seealso{
\code{\link{jaggernaut}}
}

