% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-jags.R
\name{add_jags}
\alias{add_jags}
\title{Add JAGS objects}
\usage{
add_jags(object, ...)
}
\arguments{
\item{object}{a JAGS object.}

\item{...}{additional JAGS objects to add to object.}
}
\value{
a JAGS object of the original class
}
\description{
Adds two or more JAGS object of the same class.
}

