% Generated by roxygen2 (4.0.2): do not edit by hand
\name{convergence}
\alias{convergence}
\title{Get R-hat value(s)}
\usage{
convergence(object, parm = "all", combine = TRUE, ...)
}
\arguments{
\item{object}{a JAGS object}

\item{parm}{a character vector indicating the parameters for which to calculate
the R-hat values. Either list the parmeters or use all, fixed or random.}

\item{combine}{a logical element indicating whether or not to summarise by
the maximum R-hat value.}

\item{...}{passed to and from other functions}
}
\value{
a vector, matrix or array of convergence values
}
\description{
Get R-hat value(s) for JAGS objects
}

