% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/convergence-threshold.R
\name{convergence_threshold}
\alias{convergence_threshold}
\title{R-hat threshold}
\usage{
convergence_threshold(object, ...)
}
\arguments{
\item{object}{a JAGS object.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
The R-hat threshold.
}
\description{
Gets the R-hat threshold of a JAGS object.
}
\seealso{
\code{\link{jaggernaut}}
}

