\name{data_jags.jags_simulation}
\alias{data_jags.jags_simulation}
\title{Get datasets from a JAGS simulation}
\usage{
  \method{data_jags}{jags_simulation} (object, ...)
}
\arguments{
  \item{object}{a \code{jags_simulation} object.}

  \item{...}{further arguments passed to or from other
  methods.}
}
\value{
  The datasets as a list of lists.
}
\description{
  Extracts datasets from a \code{jags_simulation} object.
}
\seealso{
  \code{\link{data_jags}} and \code{\link{jags_simulation}}
}

