% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-jags.R
\name{data_jags}
\alias{data_jags}
\title{Get dataset(s) from a JAGS object}
\usage{
data_jags(object, ...)
}
\arguments{
\item{object}{a JAGS object.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
a data.frame or list(s) of the data
}
\description{
Gets the dataset(s) from a JAGS object.
}
\seealso{
\code{\link{dataset}}
}
