\name{dataset}
\alias{dataset}
\title{Get dataset(s) from a JAGS object}
\usage{
dataset(object, ...)
}
\arguments{
  \item{object}{a JAGS object.}

  \item{...}{further arguments passed to or from other
  methods.}
}
\value{
a data.frame or list(s) of the data
}
\description{
Gets the dataset(s) from a JAGS object.
}
\seealso{
\code{\link{dataset.jags_data_model}}
}

