% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/default-model-id.R
\name{default_model_id}
\alias{default_model_id}
\title{Get default model id}
\usage{
default_model_id(object, ...)
}
\arguments{
\item{object}{a JAGS object.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
The default_model_id component of a JAGS object.
}
\description{
Get the default_model_id component of a JAGS object.
}
\seealso{
\code{\link{jaggernaut}}
}

