\name{derived_code}
\alias{derived_code}
\title{Get derived code from a JAGS object}
\usage{
derived_code(object, ...)
}
\arguments{
  \item{object}{a JAGS object.}

  \item{...}{further arguments passed to or from other
  methods.}
}
\value{
The derived_code component of object.
}
\description{
Gets the derived_code component of a JAGS object.
}
\seealso{
\code{\link{jaggernaut}}
}

