% Generated by roxygen2 (4.0.2): do not edit by hand
\name{gen_inits}
\alias{gen_inits}
\title{Get generate initial values}
\usage{
gen_inits(object, ...)
}
\arguments{
\item{object}{a JAGS object.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
The gen_inits component of object.
}
\description{
Gets the gen_inits component of of a JAGS object.
}
\seealso{
\code{\link{jaggernaut}}
}

