\docType{data}
\name{hm}
\alias{hm}
\title{House martin annual counts}
\format{A data frame with 20 rows and 2 columns}
\description{
  The house martin (\emph{Delichon urbica}) population
  annual counts from Magden (a small village in Northern
  Switzerland) collected Reto Freuler from 1990 to 2009
  from Kery & Schaub (2011 p.126). The variables are as
  follows: \itemize{ \item \code{hm} the count (integer).
  \item \code{year} the year (integer). }
}
\examples{
# data(hm)
# summary(hm)
# demo(hm)
}
\references{
  Kery M & Schaub M (2011) Bayesian Population Analysis
  using WinBUGS. Academic Press.
  (\url{http://www.vogelwarte.ch/bpa})
}
\seealso{
  \code{\link{jaggernaut}}
}
\keyword{datasets}

