\name{is.jags_analysis}
\alias{is.jags_analysis}
\title{Test for objects of class jags_analysis}
\usage{
is.jags_analysis(x)
}
\arguments{
  \item{x}{the object to be tested.}
}
\value{
a logical element indicating whether or not \code{x} is of
class \code{jags_analysis}.
}
\description{
Tests for objects of class \code{jags_analysis}.
}
\seealso{
\code{\link{jags_analysis}} and \code{\link{jaggernaut}}.
}

