% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inherits.R
\name{is.jags_model}
\alias{is.jags_model}
\title{Test for objects of class jags_model}
\usage{
is.jags_model(x)
}
\arguments{
\item{x}{the object to be tested.}
}
\value{
a logical element indicating whether or not \code{x} is of 
class \code{jags_model}.
}
\description{
Tests for objects of class \code{jags_model}.
}
\seealso{
\code{\link{jags_model}} and 
\code{\link{jaggernaut}}.
}

