\name{is.jags_sample}
\alias{is.jags_sample}
\title{Test for objects of class jags_sample}
\usage{
  is.jags_sample(x)
}
\arguments{
  \item{x}{the object to be tested.}
}
\value{
  a logical element indicating whether or not \code{x} is
  of class \code{jags_sample}.
}
\description{
  Tests for objects of class \code{jags_sample}.
}
\seealso{
  \code{\link{jags_sample}} and \code{\link{jaggernaut}}.
}

