% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is-parallel.R
\name{is_parallel}
\alias{is_parallel}
\title{Test parallel}
\usage{
is_parallel(object)
}
\arguments{
\item{object}{a JAGS object.}
}
\value{
A flag indicating whether or not the object
has parallel chains.
}
\description{
Tests JAGS object has parallel chains.
}
\seealso{
\code{\link{jaggernaut}}
}
