\docType{package}
\name{jaggernaut}
\alias{jaggernaut}
\alias{jaggernaut-package}
\alias{package-jaggernaut}
\title{Bayesian analysis with JAGS}
\description{
  Jaggernaut provides functions that facilitate Bayesian
  analyses using JAGS (Just Another Gibbs Sampler).
}
\references{
  Thorley JL (2013) Jaggernaut: An R package for Bayesian
  analysis using JAGS
}

