% Generated by roxygen2 (4.0.1): do not edit by hand
\name{jags_sample}
\alias{jags_sample}
\title{JAGS sample}
\usage{
jags_sample(chains, parm, data)
}
\arguments{
\item{chains}{\code{jagr_chains} object.}

\item{parm}{a character scalar.}

\item{data}{a data.frame.}
}
\value{
A jags_sample object.
}
\description{
An object of class \code{jags_sample}.
}
\seealso{
\code{\link{predict.jags_analysis}} and \code{\link{jaggernaut}}.
}

