\name{model}
\alias{model}
\title{Get model}
\usage{
  model(object, ...)
}
\arguments{
  \item{object}{a JAGS object.}

  \item{...}{further arguments passed to or from other
  methods.}
}
\value{
  The model component of a JAGS object.
}
\description{
  Get the model component of a JAGS object.
}
\seealso{
  \code{\link{jaggernaut}}
}

