% Generated by roxygen2 (4.0.2): do not edit by hand
\name{model_name}
\alias{model_name}
\title{Get model name}
\usage{
model_name(object, ...)
}
\arguments{
\item{object}{a JAGS object.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
The model_name component of a JAGS object.
}
\description{
Get the model_name component of a JAGS object.
}
\seealso{
\code{\link{jaggernaut}}
}

