\name{monitor<-}
\alias{monitor<-}
\title{Set monitor}
\usage{
  monitor(object) <- value
}
\arguments{
  \item{object}{a JAGS object.}

  \item{value}{a character vector of the parameters to
  monitor.}
}
\description{
  Set the monitor component of a JAGS object.
}
\seealso{
  \code{\link{jaggernaut}}
}

