% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nchains.R
\name{nchains}
\alias{nchains}
\title{Number of MCMC chains in a JAGS object}
\usage{
nchains(object)
}
\arguments{
\item{object}{a JAGS object}
}
\value{
an integer element indicating the number of MCMC chains in object
}
\description{
Gets the number of MCMC chains in a JAGS object
}

