% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmodels.R
\name{nmodels}
\alias{nmodel}
\alias{nmodels}
\title{Number of models in a JAGS object}
\usage{
nmodels(object, ...)
}
\arguments{
\item{object}{a JAGS object}

\item{...}{further arguments passed to or from other methods.}
}
\value{
an integer element indicating the number of models in object
}
\description{
Gets the number of models in a JAGS object
}

