% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsamples.R
\name{nsamples}
\alias{nsamples}
\title{Number of MCMC samples in a JAGS object}
\usage{
nsamples(object)
}
\arguments{
\item{object}{a JAGS object}
}
\value{
an integer element indicating the number of MCMC samples in object
}
\description{
Gets the number of MCMC samples in a JAGS object
}
