% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ntries.R
\name{ntries}
\alias{ntries}
\title{Get number of tries}
\usage{
ntries(object, ...)
}
\arguments{
\item{object}{a JAGS object.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
The ntries component of a JAGS object.
}
\description{
Get the ntries of a JAGS object.
}
\seealso{
\code{\link{jaggernaut}}
}

