\docType{data}
\name{p610}
\alias{p610}
\title{Point count number 610}
\format{A data frame with 146 rows and 9 columns}
\description{
  The bird point count data from the Czech republic for
  point count number 610 in 2004-2005 from Kery & Schaub
  (2011 p.157 - data courtesy of Jiri Reif).
}
\details{
  The variables are as follows: \itemize{ \item
  \code{species} the species.  \item \code{point} the point
  count number (610).  \item \code{bm} unknown.  \item
  \code{specnr} the species number .  \item \code{count1,
  count2, ..., count5} the number of individuals counted by
  occasion (1-5). }
}
\references{
  Kery M & Schaub M (2011) Bayesian Population Analysis
  using WinBUGS. Academic Press.
  (\url{http://www.vogelwarte.ch/bpa})
}
\keyword{datasets}

