% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{peregrine}
\alias{peregrine}
\title{Peregrine falcon breeding population data}
\format{A data frame with 40 rows and 4 columns}
\source{
Kery & Schaub (2011 p.64-65) courtesy of R.-J. Monneret
}
\usage{
peregrine
}
\description{
The peregrine falcon (\emph{Falco peregrinus}) population
breeding in the French Jura from 1964 to 2003.

 The variables are as follows:
\itemize{
  \item \code{Year} the year (integer).
  \item \code{Pairs} the number of adult pairs (integer).
  \item \code{R.pairs} the number of reproductive pairs (integer).
  \item \code{Eyasses} the number of fledged young (integer).
}
}
\examples{
data(peregrine)
summary(peregrine)
\dontrun{

demo(peregrine)
}
}
\references{
Kery M & Schaub M (2011) Bayesian Population Analysis
using WinBUGS. Academic Press. (\url{http://www.vogelwarte.ch/bpa})
}
\keyword{datasets}

