\docType{data}
\name{peregrine}
\alias{peregrine}
\title{Peregrine falcon breeding population}
\format{A data frame with 40 rows and 4 columns}
\description{
  The peregrine falcon (\emph{Falco peregrinus}) population
  breeding in the French Jura from 1964 to 2003 from Kery &
  Schaub (2011 p.64-65 - data courtesy of R.-J. Monneret).
  The variables are as follows: \itemize{ \item \code{Year}
  the year.  \item \code{Pairs} the number of adult pairs.
  \item \code{R.pairs} the number of reproductive pairs.
  \item \code{Eyasses} the number of fledged young. }
}
\references{
  Kery M & Schaub M (2011) Bayesian Population Analysis
  using WinBUGS. Academic Press.
  (\url{http://www.vogelwarte.ch/bpa})
}
\keyword{datasets}

