\docType{data}
\name{peregrine}
\alias{peregrine}
\title{Peregrine falcon breeding population}
\format{A data frame with 40 rows and 3 variables}
\description{
  The peregrine falcon (\emph{Falco peregrinus}) population
  breeding in the French Jura from 1964 to 2003 (data
  courtesy of R.-J. Monnerett). The variables are as
  follows: \itemize{ \item \code{Pairs} an integer vector
  of the number of adult pairs.  \item \code{R.pairs} an
  integer vector of the number of reproductive pairs.
  \item \code{Eyasses} an integer vector of the number of
  fledged young. } The data are analysed by Kery & Schaub
  (2012).
}
\references{
  Kery & Schaub (2012) Bayesian Population Analysis using
  WinBUGS. Academic Press, Oxford.
}
\keyword{datasets}

