% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.jags_analysis}
\alias{plot.jags_analysis}
\title{Plot a JAGS analysis}
\usage{
\method{plot}{jags_analysis}(x, model_id = default_model_id(x),
  parm = "fixed", ...)
}
\arguments{
\item{x}{a jags_analysis object}

\item{model_id}{a count or string specifying the jags model to select.}

\item{parm}{a character vector of the parameters to plot.}

\item{...}{arguments to pass to coda::plot.mcmc}
}
\description{
Plots the MCMC samples from a JAGS analysis model 
using the coda::plot.mcmc function.
}
\seealso{
\code{\link{jags_analysis}} and \code{\link[coda]{plot.mcmc}}
}

