% Generated by roxygen2 (4.0.2): do not edit by hand
\name{predictive_check}
\alias{predictive_check}
\title{JAGS analysis posterior predictive checks}
\usage{
predictive_check(object, parm = "discrepancy",
  model_id = default_model_id(object), derived_code = NULL,
  level = "current", estimate = "current", ...)
}
\arguments{
\item{object}{a jags_analysis object.}

\item{parm}{a character vector specifying the discrepancy
derived parameters (default = "discrepancy").}

\item{model_id}{a count or string specifying the jags model to select.}

\item{derived_code}{a character scalar defining a block in the
JAGS dialect of  the BUGS language that defines the discrepancy(s).}

\item{level}{a numeric scalar specifying the significance level or a character
scalar specifying which mode the level should be taken from. By default the
level is as currently specified by \code{opts_jagr} in the global options.}

\item{estimate}{a character scalar indicating whether the point estimate
should be the "mean" or the "median". By default the estimate is as
currently defined by \code{opts_jagr} in the global options.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
a coef table of the posterior predictive check p-values(s)
}
\description{
Calculate posterior predictive check p-value for a JAGS analysis
}
\examples{
\dontrun{
model1 <- jags_model("model {
alpha ~ dnorm(0, 50^-2)
beta ~ dnorm(0, 10^-2)
sigma ~ dunif(0, 10)

for(i in 1:length(Volume)) {
  eMu[i] <- alpha + beta * Girth[i]
  Volume[i] ~ dnorm(eMu[i], sigma^-2)
}
}",
derived_code = "data {
for(i in 1:length(Volume)) {
  prediction[i] <- alpha + beta * Girth[i]

  simulated[i] ~ dnorm(prediction[i], sigma^-2)

  D_observed[i] <- log(dnorm(Volume[i], prediction[i], sigma^-2))
  D_simulated[i] <- log(dnorm(simulated[i], prediction[i], sigma^-2))
}
discrepancy <- sum(D_observed) - sum(D_simulated)

}",
select_data = c("Volume", "Girth*"))

data(trees)
analysis1 <- jags_analysis(model1, data = trees)
predictive_check(analysis1)
}
}
\seealso{
\code{\link{jags_model}}, \code{\link{jags_analysis}}
,\code{\link{predict.jags_analysis}} and \code{\link{coef.jags_analysis}}
}

