\name{print.janalysis}
\alias{print.janalysis}
\title{Print a JAGS analysis}
\usage{
  \method{print}{janalysis} (x, ...)
}
\arguments{
  \item{x}{a janalysis object to print}

  \item{...}{further arguments pass to or from other
  methods.}
}
\description{
  Prints a JAGS analysis
}
\seealso{
  \code{\link{jaggernaut}}
}

