\name{print.summary_janalysis}
\alias{print.summary_janalysis}
\title{Print summary of JAGS analysis (janalysis) object}
\usage{
  print.summary_janalysis(x)
}
\arguments{
  \item{x}{the summary of a JAGS analysis (janalysis)
  object}
}
\description{
  Prints a summary of a JAGS analysis (janalysis) object
}

