% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select-data.R
\name{select_data<-}
\alias{select_data<-}
\title{Set select_data}
\usage{
select_data(object) <- value
}
\arguments{
\item{object}{a JAGS object.}

\item{value}{a character vector of the variables to select or NULL.}
}
\description{
Set the select_data component of a JAGS object.
}
\seealso{
\code{\link{jaggernaut}}
}

