% Generated by roxygen2 (4.0.2): do not edit by hand
\name{subset.jags_analysis}
\alias{subset.jags_analysis}
\title{Subset a JAGS analysis}
\usage{
\method{subset}{jags_analysis}(x, model_id = 1:nmodels(x), ...)
}
\arguments{
\item{x}{a jags_analysis x.}

\item{model_id}{an integer or character vector specifying the jags models to select.}

\item{...}{other arguments passed to generic function.}
}
\value{
the subsetted jags_analysis x
}
\description{
Subset a JAGS analysis x.
}
\seealso{
\code{\link{subset}} and \code{\link{jags_analysis}}
}

