% Generated by roxygen2 (4.0.2): do not edit by hand
\name{subset.jags_model}
\alias{subset.jags_model}
\title{Subset a JAGS model}
\usage{
\method{subset}{jags_model}(x, ...)
}
\arguments{
\item{x}{a jags_model x.}

\item{...}{other arguments passed to generic function.}

\item{model}{an integer or character vector specifying the jags models to select.}
}
\value{
the subsetted jags_analysis x
}
\description{
Subset a JAGS model x.
}
\seealso{
\code{\link{subset}} and \code{\link{jags_analysis}}
}

