% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/subset.R
\name{subset.jags_model}
\alias{subset.jags_model}
\title{Subset a JAGS model}
\usage{
\method{subset}{jags_model}(x, model_id = default_model_id(x), ...)
}
\arguments{
\item{x}{a jags_model x.}

\item{model_id}{an integer or character vector specifying the jags models to select.}

\item{...}{other arguments passed to generic function.}
}
\value{
the subsetted jags_analysis x
}
\description{
Subset a JAGS model x.
}
\seealso{
\code{\link{subset}} and \code{\link{jags_analysis}}
}

