\name{subset.jags_model}
\alias{subset.jags_model}
\title{Subset a JAGS model}
\usage{
  \method{subset}{jags_model} (x, model_number, ...)
}
\arguments{
  \item{x}{a jags_model x.}

  \item{model_number}{an integer element indicating the
  model(s) to select.}

  \item{...}{other arguments passed to generic function.}
}
\value{
  the subsetted jags_analysis x
}
\description{
  Subset a JAGS model x.
}
\seealso{
  \code{\link{subset}} and \code{\link{jags_analysis}}
}

