\name{subset_jags}
\alias{subset_jags}
\title{Subset a JAGS object}
\usage{
  subset_jags(object, ...)
}
\arguments{
  \item{object}{a JAGS object.}

  \item{...}{other arguments.}
}
\value{
  an object of the same JAGS class.
}
\description{
  Subsets a JAGS object.
}
\seealso{
  \code{\link{jaggernaut}}
}

