\name{summary.janalysis}
\alias{summary.janalysis}
\title{Summary of JAGS analysis (janalysis) object}
\usage{
  summary.janalysis(object, parameters = NULL)
}
\arguments{
  \item{object}{a JAGS analysis (janalysis) object to
  summarise}
}
\description{
  Produces a summary of a JAGS analysis (janalysis) object
}

