\name{summary.janalysis}
\alias{summary.janalysis}
\title{Summary of JAGS analysis}
\usage{
  \method{summary}{janalysis} (object, ...)
}
\arguments{
  \item{object}{a janalysis object to summarise}

  \item{...}{further arguments to pass to or from other
  methods.}
}
\value{
  a jsummary_analysis object
}
\description{
  Produces a summary of a JAGS analysis
}
\seealso{
  \code{\link{analysis}}, \code{\link{tits}}
}

